requirements.txt
mss>=9.0.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
setproctitle>=1.3.0
pytesseract>=0.3.10
# 키워드 OCR 엔진 (선택)
easyocr>=1.7.0
rapidocr-onnxruntime>=1.3.0
